Update README.md
Although Kolosal AI is entirely open and can be used for commercial purposes, the engine can not be distributed for commercial purposes. If you are planning to sell this engine for commercial, you'd need to get the enterprise subscription of Kolosal AI, or implement other engine entirely on Kolosal AI to distribute without distributing this engine.